IOS Network: Split ncd/manage code into its own source files
