+++ Source/Core/Core/CMakeLists.txt
@@ -146,7 +146,6 @@ set(SRCS	ActionReplay.cpp
 			IOS/ES/Formats.cpp
 			IOS/FS/FileIO.cpp
 			IOS/FS/FS.cpp
-			IOS/Network/Config.cpp
 			IOS/Network/ICMPLin.cpp
 			IOS/Network/MACUtils.cpp
 			IOS/Network/Net.cpp
@@ -156,6 +155,8 @@ set(SRCS	ActionReplay.cpp
 			IOS/Network/KD/NetKDRequest.cpp
 			IOS/Network/KD/NetKDTime.cpp
 			IOS/Network/KD/NWC24Config.cpp
+			IOS/Network/NCD/Config.cpp
+			IOS/Network/NCD/Manage.cpp
 			IOS/SDIO/SDIOSlot0.cpp
 			IOS/STM/STM.cpp
 			IOS/USB/Common.cpp
