+++ Source/Core/Core/Core.vcxproj
@@ -178,7 +178,6 @@
     <ClCompile Include="IOS\ES\Formats.cpp" />
     <ClCompile Include="IOS\FS\FileIO.cpp" />
     <ClCompile Include="IOS\FS\FS.cpp" />
-    <ClCompile Include="IOS\Network\Config.cpp" />
     <ClCompile Include="IOS\Network\ICMPLin.cpp" />
     <ClCompile Include="IOS\Network\MACUtils.cpp" />
     <ClCompile Include="IOS\Network\Net.cpp" />
@@ -188,6 +187,8 @@
     <ClCompile Include="IOS\Network\KD\NetKDRequest.cpp" />
     <ClCompile Include="IOS\Network\KD\NetKDTime.cpp" />
     <ClCompile Include="IOS\Network\KD\NWC24Config.cpp" />
+    <ClCompile Include="IOS\Network\NCD\Config.cpp" />
+    <ClCompile Include="IOS\Network\NCD\Manage.cpp" />
     <ClCompile Include="IOS\SDIO\SDIOSlot0.cpp" />
     <ClCompile Include="IOS\STM\STM.cpp" />
     <ClCompile Include="IOS\USB\Common.cpp" />
@@ -419,7 +420,6 @@
     <ClInclude Include="IOS\ES\Formats.h" />
     <ClInclude Include="IOS\FS\FileIO.h" />
     <ClInclude Include="IOS\FS\FS.h" />
-    <ClInclude Include="IOS\Network\Config.h" />
     <ClInclude Include="IOS\Network\ICMPLin.h" />
     <ClInclude Include="IOS\Network\ICMP.h" />
     <ClInclude Include="IOS\Network\MACUtils.h" />
@@ -430,6 +430,8 @@
     <ClInclude Include="IOS\Network\KD\NetKDRequest.h" />
     <ClInclude Include="IOS\Network\KD\NetKDTime.h" />
     <ClInclude Include="IOS\Network\KD\NWC24Config.h" />
+    <ClInclude Include="IOS\Network\NCD\Config.h" />
+    <ClInclude Include="IOS\Network\NCD\Manage.h" />
     <ClInclude Include="IOS\SDIO\SDIOSlot0.h" />
     <ClInclude Include="IOS\STM\STM.h" />
     <ClInclude Include="IOS\USB\Common.h" />
