+++ Source/Core/Core/IOS/Network/NCD/Config.cpp
@@ -2,7 +2,7 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
-#include "Core/IOS/Network/Config.h"
+#include "Core/IOS/Network/NCD/Config.h"
 
 #include <cstring>
