+++ Source/Core/Core/IOS/Network/Net.cpp
@@ -24,79 +24,6 @@ namespace HLE
 {
 namespace Device
 {
-// **********************************************************************************
-// Handle /dev/net/ncd/manage requests
-NetNCDManage::NetNCDManage(u32 device_id, const std::string& device_name)
-    : Device(device_id, device_name)
-{
-}
-
-IPCCommandResult NetNCDManage::IOCtlV(const IOCtlVRequest& request)
-{
-  s32 return_value = IPC_SUCCESS;
-  u32 common_result = 0;
-  u32 common_vector = 0;
-
-  switch (request.request)
-  {
-  case IOCTLV_NCD_LOCKWIRELESSDRIVER:
-    break;
-
-  case IOCTLV_NCD_UNLOCKWIRELESSDRIVER:
-    // Memory::Read_U32(request.in_vectors.at(0).address);
-    break;
-
-  case IOCTLV_NCD_GETCONFIG:
-    INFO_LOG(IOS_NET, "NET_NCD_MANAGE: IOCTLV_NCD_GETCONFIG");
-    config.WriteToMem(request.io_vectors.at(0).address);
-    common_vector = 1;
-    break;
-
-  case IOCTLV_NCD_SETCONFIG:
-    INFO_LOG(IOS_NET, "NET_NCD_MANAGE: IOCTLV_NCD_SETCONFIG");
-    config.ReadFromMem(request.in_vectors.at(0).address);
-    break;
-
-  case IOCTLV_NCD_READCONFIG:
-    INFO_LOG(IOS_NET, "NET_NCD_MANAGE: IOCTLV_NCD_READCONFIG");
-    config.ReadConfig();
-    config.WriteToMem(request.io_vectors.at(0).address);
-    break;
-
-  case IOCTLV_NCD_WRITECONFIG:
-    INFO_LOG(IOS_NET, "NET_NCD_MANAGE: IOCTLV_NCD_WRITECONFIG");
-    config.ReadFromMem(request.in_vectors.at(0).address);
-    config.WriteConfig();
-    break;
-
-  case IOCTLV_NCD_GETLINKSTATUS:
-    INFO_LOG(IOS_NET, "NET_NCD_MANAGE: IOCTLV_NCD_GETLINKSTATUS");
-    // Always connected
-    Memory::Write_U32(Net::ConnectionSettings::LINK_WIRED, request.io_vectors.at(0).address + 4);
-    break;
-
-  case IOCTLV_NCD_GETWIRELESSMACADDRESS:
-    INFO_LOG(IOS_NET, "NET_NCD_MANAGE: IOCTLV_NCD_GETWIRELESSMACADDRESS");
-
-    u8 address[Common::MAC_ADDRESS_SIZE];
-    IOS::Net::GetMACAddress(address);
-    Memory::CopyToEmu(request.io_vectors.at(1).address, address, sizeof(address));
-    break;
-
-  default:
-    INFO_LOG(IOS_NET, "NET_NCD_MANAGE IOCtlV: %#x", request.request);
-    break;
-  }
-
-  Memory::Write_U32(common_result, request.io_vectors.at(common_vector).address);
-  if (common_vector == 1)
-  {
-    Memory::Write_U32(common_result, request.io_vectors.at(common_vector).address + 4);
-  }
-  return GetDefaultReply(return_value);
-}
-
-// **********************************************************************************
 // Handle /dev/net/wd/command requests
 NetWDCommand::NetWDCommand(u32 device_id, const std::string& device_name)
     : Device(device_id, device_name)
