+++ Source/Core/Core/IOS/Network/Net.h
@@ -9,11 +9,6 @@
 #include "Common/CommonTypes.h"
 #include "Core/IOS/Device.h"
 #include "Core/IOS/IPC.h"
-#include "Core/IOS/Network/Config.h"
-
-#ifdef _WIN32
-#include <ws2tcpip.h>
-#endif
 
 namespace IOS
 {
@@ -22,32 +17,6 @@ namespace HLE
 // TODO: split this up.
 namespace Device
 {
-// **********************************************************************************
-// Interface for reading and changing network configuration (probably some other stuff as well)
-class NetNCDManage : public Device
-{
-public:
-  NetNCDManage(u32 device_id, const std::string& device_name);
-
-  IPCCommandResult IOCtlV(const IOCtlVRequest& request) override;
-
-private:
-  enum
-  {
-    IOCTLV_NCD_LOCKWIRELESSDRIVER = 0x1,    // NCDLockWirelessDriver
-    IOCTLV_NCD_UNLOCKWIRELESSDRIVER = 0x2,  // NCDUnlockWirelessDriver
-    IOCTLV_NCD_GETCONFIG = 0x3,             // NCDiGetConfig
-    IOCTLV_NCD_SETCONFIG = 0x4,             // NCDiSetConfig
-    IOCTLV_NCD_READCONFIG = 0x5,
-    IOCTLV_NCD_WRITECONFIG = 0x6,
-    IOCTLV_NCD_GETLINKSTATUS = 0x7,          // NCDGetLinkStatus
-    IOCTLV_NCD_GETWIRELESSMACADDRESS = 0x8,  // NCDGetWirelessMacAddress
-  };
-
-  Net::WiiNetConfig config;
-};
-
-//////////////////////////////////////////////////////////////////////////
 class NetWDCommand : public Device
 {
 public:
